:recycle: Updated NotificationCenter components
