+++ packages/bell/src/components/notification-center/NotificationCenter.tsx
@@ -1,22 +1,28 @@
-import React from 'react';
+import React, { useContext } from 'react';
 import { IMessage } from '@novu/shared';
 
-import { App } from './components/App';
+import { NovuContext } from '../../store/novu-provider.context';
+import { NotificationCenterContext } from '../../store/notification-center.context';
+import { AppContent } from './components';
 
 export interface INotificationCenterProps {
   onUrlChange?: (url: string) => void;
   onNotificationClick?: (notification: IMessage) => void;
   onUnseenCountChanged?: (unseenCount: number) => void;
-  bell?: JSX.Element | Element;
 }
 
 export function NotificationCenter(props: INotificationCenterProps) {
+  const { applicationIdentifier } = useContext(NovuContext);
+
   return (
-    <App
-      onNotificationClick={props.onNotificationClick}
-      onUnseenCountChanged={props.onUnseenCountChanged}
-      onUrlChange={props.onUrlChange}
-      bell={props.bell}
-    />
+    <NotificationCenterContext.Provider
+      value={{
+        sendUrlChange: props.onUrlChange,
+        sendNotificationClick: props.onNotificationClick,
+        onUnseenCountChanged: props.onUnseenCountChanged,
+        isLoading: !applicationIdentifier,
+      }}>
+      <AppContent />
+    </NotificationCenterContext.Provider>
   );
 }
